requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.2.0
colorama==0.4.6
cryptography==42.0.5
Flask==3.1.0
frozenlist==1.4.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.0.5
mysql-connector-python==9.3.0
pillow==10.4.0
pycparser==2.22
pyfingerprint==1.5
PyJWT==2.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
pyserial==3.5
PyYAML==6.0.2
qrcode==8.2
requests==2.32.3
twilio==9.2.4
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.2.2
user-agents==2.2.0
Werkzeug==3.1.3
yarl==1.9.4
